another stab at numpy 2 support

+++ mdtraj/core/trajectory.py
@@ -874,7 +874,7 @@ class Trajectory:
             if the Trajectory contains no unitcell information.
         """
         if self.unitcell_lengths is not None:
-            return np.array(list(map(np.linalg.det, self.unitcell_vectors)))
+            return np.array(list(map(np.linalg.det, self.unitcell_vectors)), dtype=np.float64)
         else:
             return None
